Updated UI
Pause Menu added to the UI, includes resume, restart, and quit buttons. Added ButtonFunctions object with ButtonFunctions Script to the UI to provide functionality to the above Buttons. GameManager Now includes functionality to pause and unpause game state when the escape key is pressed. Code in GunBase altered on line 25 to check if the game is paused to not fire.